requirements.txt
django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0
requests>=2.31
